Updated up, added more features

+++ app/controllers/login_c.js
@@ -23,6 +23,7 @@ function loginController($scope, $rootScope,login_service)
 	$scope.login = function(account){
 		var post = "&username=" + account.username;
 		post += "&password=" + account.password;
+		post += "&member_type=" +soleinsider.member_type;
 
 		login_service.login(post).then(function (data) {
 
@@ -41,26 +42,50 @@ function loginController($scope, $rootScope,login_service)
 
 	};
 
+	$scope.validateEmail = function(email){
+    	var re = /\S+@\S+\.\S+/;
+    	return re.test(email);
+	};
+
+	$scope.validateAccount = function(account){
+		if (account.username.length == 0 || account.password.length == 0 || account.phone_number.length == 0 || $scope.validateEmail(account.username) == false){
+			return false;
+		} else{
+			return true;
+		}
+	};
+
 	$scope.register = function(newaccount){
 
-		var post = "&username=" + newaccount.username;
-			post += "&password=" + newaccount.password;
-			post += "&phone=" + newaccount.phone_number;
-			post += "&carrier=" + newaccount.carrier;
-			post += "&member_type" +soleinsider.member_type;
+		var validated = $scope.validateAccount(newaccount);
 
-		login_service.createAccount(post).then(function (data) {
+		if(!validated){
+			$("html, body").animate({ scrollTop: 0 }, "slow");
 			$scope.showConfirmation = true;
-			if (data !== "false" && data !== false && data.length !== 0){
-				localStorage.setItem("username", newaccount.username);
-				localStorage.setItem("member_id", data);
-				$scope.confirmation = "Your account have been created";
-			} else{
-				$scope.confirmation = "This username is already in use";
-			}
-		}, function (err) {
-			window.console.log(err);
-		});
+			$scope.confirmation = "Incorrect information";
+			return;
+		} else{
+
+			var post = "&username=" + newaccount.username;
+				post += "&password=" + newaccount.password;
+				post += "&phone=" + newaccount.phone_number;
+				post += "&carrier=" + newaccount.carrier;
+				post += "&member_type=" +soleinsider.member_type;
+
+			login_service.createAccount(post).then(function (data) {
+				$scope.showConfirmation = true;
+				if (data !== "false" && data !== false && data.length !== 0){
+					localStorage.setItem("username", newaccount.username);
+					localStorage.setItem("member_id", data);
+					$scope.confirmation = "Your account have been created";
+				} else{
+					$scope.confirmation = "This username is already in use";
+				}
+			}, function (err) {
+				window.console.log(err);
+			});
+
+		}
 
 
 	}
